Fix crash with large jitter.
When a large jitter is passed (close to the size of the file), a chunk
can be found that spans the whole file from start to end. This will
obviously have the same map as the whole file (since they are one and
the same) and because nothing checks that the results found are smaller
than the starting test case, it will end up in RESULTS. If this is the
only result, the size of the chunk will never be less than best_size, so
best_result remains None and gets subscripted as chunks[0] soon after,
in normalize_chunks.

This commit filters out chunks that have the same size as the starting
size to prevent this.